# Conflicts:
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AutoPickup.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RobotChildren/NessieTeleOp.java

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/MecanumDrive.java
@@ -631,7 +631,7 @@ public final class MecanumDrive
         rightBack.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
         power *= right;
-        int targetPos = rightFront.getCurrentPosition();
+        double targetPos = rightFront.getCurrentPosition();
         targetPos += target * PARAMS.inPerTick;
 
         if ((Math.abs(rightFront.getCurrentPosition()) <= targetPos))
